using the multiprocessing queue instead of the standerd one to demonstrate the use of IPC

+++ thumbnail_maker.py
@@ -6,20 +6,20 @@ from urllib.parse import urlparse
 from urllib.request import urlretrieve
 from queue import Queue
 from threading import Thread
-import multiprocessing 
+import multiprocessing
 
 import PIL
 from PIL import Image
 
 FORMAT = "[%(threadName)s, %(asctime)s, %(levelname)s] %(message)s"
-logging.basicConfig(filename='logfile.log', level=logging.DEBUG, format=FORMAT, force=True)
+logging.basicConfig(filename='logfile.log', level=logging.DEBUG, format=FORMAT)
 
 class ThumbnailMakerService(object):
     def __init__(self, home_dir='.'):
         self.home_dir = home_dir
         self.input_dir = self.home_dir + os.path.sep + 'incoming'
         self.output_dir = self.home_dir + os.path.sep + 'outgoing'
-        self.img_list = []
+        self.img_queue = multiprocessing.JoinableQueue()
 
     def download_image(self, dl_queue):
         while not dl_queue.empty():
@@ -27,62 +27,75 @@ class ThumbnailMakerService(object):
                 url = dl_queue.get(block=False)
                 # download each image and save to the input dir
                 img_filename = urlparse(url).path.split('/')[-1]
-
                 urlretrieve(url, self.input_dir + os.path.sep + img_filename)
-                self.img_list.append(img_filename)
+                self.img_queue.put(img_filename)
+
                 dl_queue.task_done()
-
             except Queue.Empty:
-                logging.info("Queue empty")
+                logging.info('Queue empty')
 
-    def resize_image(self, filename):
+    def perform_resizing(self):
+        # validate inputs
+        os.makedirs(self.output_dir, exist_ok=True)
+
+        logging.info("beginning image resizing")
         target_sizes = [32, 64, 200]
-        logging.info("resizing image {}".format(filename))
-        orig_img = Image.open(self.input_dir + os.path.sep + filename)
-        for basewidth in target_sizes:
-            img = orig_img
-            # calculate target height of the resized image to maintain the aspect ratio
-            wpercent = (basewidth / float(img.size[0]))
-            hsize = int((float(img.size[1]) * float(wpercent)))
-            # perform resizing
-            img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
-            
-            # save the resized image to the output dir with a modified file name 
-            new_filename = os.path.splitext(filename)[0] + \
-                '_' + str(basewidth) + os.path.splitext(filename)[1]
-            img.save(self.output_dir + os.path.sep + new_filename)
+        num_images = len(os.listdir(self.input_dir))
 
-        os.remove(self.input_dir + os.path.sep + filename)
-        logging.info("done resizing image {}".format(filename))
+        start = time.perf_counter()
+        while True:
+            filename = self.img_queue.get()
+            if filename:
+                logging.info("resizing image {}".format(filename))
+                orig_img = Image.open(self.input_dir + os.path.sep + filename)
+                for basewidth in target_sizes:
+                    img = orig_img
+                    # calculate target height of the resized image to maintain the aspect ratio
+                    wpercent = (basewidth / float(img.size[0]))
+                    hsize = int((float(img.size[1]) * float(wpercent)))
+                    # perform resizing
+                    img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
 
+                    # save the resized image to the output dir with a modified file name
+                    new_filename = os.path.splitext(filename)[0] + \
+                        '_' + str(basewidth) + os.path.splitext(filename)[1]
+                    img.save(self.output_dir + os.path.sep + new_filename)
+
+                os.remove(self.input_dir + os.path.sep + filename)
+                logging.info("done resizing image {}".format(filename))
+                self.img_queue.task_done()
+            else:
+                self.img_queue.task_done()
+                break
+        end = time.perf_counter()
+
+        logging.info("created {} thumbnails in {} seconds".format(num_images, end - start))
 
     def make_thumbnails(self, img_url_list):
         logging.info("START make_thumbnails")
-        pool= multiprocessing.Pool()
 
         start = time.perf_counter()
-
+        
         dl_queue = Queue()
 
         for img_url in img_url_list:
             dl_queue.put(img_url)
-
 
         num_dl_threads = 4
         for _ in range(num_dl_threads):
             t = Thread(target=self.download_image, args=(dl_queue,))
             t.start()
 
+        num_processes = multiprocessing.cpu_count()
+        for _ in range (num_processes):
+
+            p = multiprocessing.Process(target=self.perform_resizing)
+            p.start()
+
         dl_queue.join()
-        
-        start_resize = time.perf_counter()
-        pool.map(self.resize_image, self.img_list)
-        end_resize = time.perf_counter()
-        pool.close()
-        pool.join()
+        # add poision apple per process
+        for _ in range (num_processes):
+            self.img_queue.put(None)
 
         end = time.perf_counter()
-        logging.info("created {} thumbnails in {} seconds".format(len(self.img_list), end_resize - start_resize))
-
         logging.info("END make_thumbnails in {} seconds".format(end - start))
-
